Test: fiddle with the arbitrary modules
