+++ test/Icicle/Test/Source/PrettyParse.hs
@@ -9,7 +9,7 @@ import           Icicle.Test.Arbitrary
 import           Icicle.Internal.Pretty
 import           Icicle.Source.Parser
 
-import           Icicle.Test.Source.Arbitrary
+import           Icicle.Test.Arbitrary
 
 import           P
