+++ test/Icicle/Test/Source/Progress.hs
@@ -18,8 +18,7 @@ import           Icicle.Source.Transform.Desugar
 import qualified Icicle.Common.Base          as CB
 import qualified Icicle.Common.Type          as CT
 
-import           Icicle.Test.Source.Arbitrary
-import           Icicle.Test.Core.Arbitrary ()
+import           Icicle.Test.Arbitrary
 
 import           P
